Refactor: Cleaned code, improved folder structure, and applied best practices
