+++ app/layout.js
@@ -1,36 +1,15 @@
-import { Footer, Layout, Navbar } from "nextra-theme-docs";
+import { Layout } from "nextra-theme-docs";
 import "nextra-theme-docs/style.css";
 import { Head } from "nextra/components";
 import { getPageMap } from "nextra/page-map";
 import "./globals.css";
+import Navbar from "@/components/layout/Navbar";
+import Footer from "@/components/layout/Footer";
 
 export const metadata = {
   title: "thrivext",
   description: "Notes in Bangla covering various modern web technologies.",
 };
-
-const CustomNavbar = () => {
-  return (
-    <Navbar
-      logo={
-        <div className="flex items-center  w-full justify-between relative">
-          <span className="text-3xl font-bold">Thrivext</span>
-        </div>
-      }
-      projectLink="https://github.com/Takib-Ahmed/thrivext"
-    />
-  );
-};
-
-const footer = (
-  <div className="nx-text-gray-500">
-    <Footer>
-      <h1 className="nx-text-center">
-        copyright {new Date().getFullYear()} © thrivext.
-      </h1>
-    </Footer>
-  </div>
-);
 
 export default async function RootLayout({ children }) {
   return (
@@ -44,15 +23,15 @@ export default async function RootLayout({ children }) {
       </Head>
       <body suppressHydrationWarning>
         <Layout
-          navbar={<CustomNavbar />}
+          navbar={<Navbar />}
           pageMap={await getPageMap()}
           sidebar={{
             defaultMenuCollapseLevel: 1,
           }}
           docsRepositoryBase="https://github.com/Takib-Ahmed/thrivext/tree/development"
-          footer={footer}
+          footer={<Footer />}
         >
-          {children}
+          <main>{children}</main>
         </Layout>
       </body>
     </html>
