+++ public/images/TypeScript/TypeScript_type.png
